Fix video not shown correctly on landscape orientation
The video view behavior has changed for aspectFit scale, and now automatically stretches itself to cover the size of its container when rotating, while maintaining the aspect ratio. Thus, the only needed part from the widgets side is to have the video view cover the whole view bounds, center it, and change the scale to aspectFit.

MUIC-320