Abstract CoreSDK in Widgets to introduce unit tests
This PR aims to start process of abstraction core SDK in order make Widgets testable. This is the first step: narrow down imports of core SDK and use CoreSDKClient namespace instead. It is necessary to be done soon since changes are impactful and may introduce merge conflicts with ongoing development of features if addressed later.

MOB-1120