Add initial implementation of VoiceOver
The goal of this PR is to enable VoiceOver on Widgets SDK. Currently this is only WIP for call and chat screens. Note that localization is not taken into account, it will be addressed in upcoming PR(s) along with snapshot tests. Feel free to add ideas and suggestions how it can be improved. Since this PR has a lot of changes I think is is reasonable to merge it as is and add proper localization later.

MOB-1140