Add support for incrementing project version through fastlane
To make it easier to upgrade versions of the widgets, `fastlane` is
added. With it, a lane is created to update the version in the
Info.plists. Also, an action is added that goes into the
`StaticValues.swift` file and modifies the value of it. This is to
avoid having logs with incorrect versions in case integrators modify
our Info.plist files somehow. This action is then used in the
`increment_project_version` lane as well.

MOB-1375