Add accessibility identifier for audio upgrades
In order to detect if an engagement has been upgraded to audio, new
accessibility identifiers have been added to the buttons in the audio
call view, and the alert that appears when the operator requests an
audio call from the visitor. For the audio call view, the
accessibility identifier is only set when the button is enabled.

MOB-1400

Fix