Add Bitrise workflow to upgrade project version
A new lane is added that increments the project version using the
already available lane. Also, a new workflow in `bitrise.yml` is
added that uses this lane along with an EnvVar that one can control
by sending `major`, `minor`, or `patch`. Some changes to naming were
done, as the previous names were not appropriate with the current
changes.

MOB-1507