Update `bitrise.yml` release workflow
Two additional workflows are going to be executed when releasing the
widgets. The first one will create a new podspec with the new version
of the widgets. This is done by executing the required workflow in the
`glia-ios-podspecs` project. The second is part of this same project,
and that will update the current project version to the next minor
version. This ensures the repository is always ready to release
without any additional PRs at the moment of release.

MOB-1517