Update Bitrise deployment workflow
The Bitrise workflow has been updated to receive a `NEW_VERSION`
environment variable, in order to be synced to what we do in the Core
SDK. Apart from that, the increment_project_version is updated so that
it doesn’t have unnecessary steps. This time it will succeed always.

MOB-1589