Snapshot tests fixes
Since tests were broken previously by deleting applying frame to tested view, I returned it back. If the frame for the tested view is not set and the reference for the test already exists, test will pass, but this is not correct.
Also fixed mocks for chat screen tests.
Also fixed bottom padding between buttons and safeAreaLayoutGuide on the survey screen