Add possibility to increment the version through GitHub Actions
Add the same work that is now present in all Android repositories,
where you can run a GitHub Action that lets you choose the increment
type, and Bitrise will create a PR that increments the version number.
Also, the same approach is used as in Android where the variable sent
through GitHub is replacing the envvar in Bitrise. This is done in a
script step in Bitrise, and is the approach recommended by Bitrise
themselves.

MOB-1625