Create example integration app
Currently the integration process took more time than it should.
According to recent reports over 20 minutes. To reduce the time for
initial setup, a sample app was created to demonstrate the primary use
of Glia's WidgetSDK. The sample app allows to start an engagement with
text, audio, and video.

MOB-1638