Fix delivered status text being modified through the theme
Before, setting the delivered status text wouldn't do anything
because the widgets got the text from the localized strings instead.
Now the widgets are getting the text from the theme sent as a
parameter for the start. If the text is not overriden, then it is
still using `Delivered` as status.

MOB-1704