Add confirmation view to secure messaging
Add all necessary files for showing and for data-driven approach. Also
make changes so that this view can be interchanged with the welcome
screen. We need to do something like that anyways because the secure
conversations can start with welcome screen or with chat transcript
screen. Also move the welcome files to their own folder so that we can
have a folder per screen within secure conversations.

MOB-1728