Show alert for unavailable message center
When the message center is unavailable, an alert is shown which is
different from what is currently in place. The alert needs to have
transparent background and you need to be able to interact with the
view behind. All our alerts are currently presented modally, which
doesn’t allow a straightforward way to pass touches to the view
behind. Thus, this alert is added as a subview instead of modally.
Instead of trying to use the current method, I created a new one so
that there’s a clear distinction between both. This can be discussed.

MOB-1724