Add accessibility to welcome screen
Because the accessibility labels and hints need to be localized, all
strings from the welcome screen have been also localized. Labels and
hints have been added where needed. The commit seems big but the
`renderProps()` method in the welcome view had to be moved somewhere
else to avoid a SwiftLint issue.

MOB-1710