Add accessibility and localization for confirmation screen
Just as in welcome screen, it is better to add accessibility together
with localization so that we can use localized accessibility labels
and hints. At some point someone moved the welcome styles to their own
files, so that is also done here because it makes it easier to
navigate each specific style.

MOB-1890