Fix labels in confirmation screen displaying large text incorrectly
When enabling large text in accessibility, the labels would cut off
the text and just show one line because they were wrongly constrained
by height. However when removing the constraint, the screen would look
bad because the stack view was set to fill proportionally. This was
changed to `fill` to allow removing the constraints and keeping the
requested design in Figma.

MOB-1890