Remove the back button on secure conversations chat transcript
The current logic correctly identifies when to show the secure
conversations title or the chat title. This logic is encapsulated into
its own function so that it can be used to determine back button
behavior. The secure chat transcript header style is the same as the
regular chat style but with the back button set as `nil`.

MOB-1888