Remove the back button on secure conversations chat transcript
The current logic correctly identifies when to show the secure
conversations title or the chat title. This logic is encapsulated
into its own function so that it can be used to determine back button
behavior. The whole props for the header are regenerated every time
the new renderProps method (previously renderTitle) is called, so that
it correctly identifies the different chat models and its actions.

MOB-1888