Keep display on during audio and video calls
Currently during audio or video call the screen would turn off after some time if there is no interaction with the screen. This PR fixes this issue and keeps the screen on until the end of the call. In addition, if user decides to hold the phone to their ear, the proximity sensor is activated, which will dim the screen and disable interaction until phone has been removed from close contact.

MOB-1661