Add support for showing operator image in chat transcript
In case there is no `SalemoveOperator`, that is, when there is no
engaged operator because the engagement hasn’t started yet, then the
chat sender information is used to figure out the operator for the
chat message. The main goal is to have it show a picture on secure
transcripts, where right now the picture is completely missing.

MOB-2017