Add support for welcome messages in secure transcript screen
Some work is done on the secure transcript view model screen and some
on the chat view model screen because of MSG-483. This bug means that
receiving and displaying the welcome message is done when the secure
transcript view model is active, but the handling of the response is
done when the chat view model is active.

MOB-2116