Improve response cards
currently response cards behaved differently on iOS compared to web. In
web response card would collapse upon a click, and would show the choice
as a normal text. on iOS the response card remained open and just highlighted
the selection. This PR fixes it and iOS response cards will behave like in web.