Fix attachment not being sent from welcome screen while on engagement
When sending an image from the welcome screen, we need to know if an
engagement is ongoing or not. If an attachment is sent to the secure
messaging endpoint while on an engagement, it doesn't arrive. Thus,
we check for the current engagement every time a file is uploaded.
This cannot be done only on the init of the environment, because the
status of the ongoing engagement can change between the moment that
the screen is shown and files are uploaded.

MOB-2164