Avoid presenting survey in welcome and confirmation screen
The issue happened because the coordinator called `end()` for all
screens, even if they are welcome and confirmation, which should never
be in a situation where they need to present a survey. Thus, a new
enum was created called `SurveyPresentation`, which determines if the
screen needs to present a survey. This is an enum and not a boolean
because it's an associated value in an enum, so it's more descriptive.

MOB-2163