Add support for push notifications timing
Receiving a push notification while in the app would make the push
handler be executed because there was no way to differentiate between
pushes that came while in the app vs pushes in the background. Thanks
to the `timing` property added to `Push`, we can now only execute
the required stuff when the push notification came while the app was
on the background and the visitor accepts it.

Also, support for the new type `queuedMessage` has been added.

MOB-2177
MOB-2161