Simulate welcome messages arrive after engagement start event
The backend sends us messages in the incorrect order. However, product
still wants us to show them correctly. Thus, what is done is the
following:

1. Have socket only listen for visitor and system messages.
2. When the engagement starts, have the chat transcript downloaded.
3. Determine which operator messages are new, and add those to the
messages section, not the history section.
4. Restart the socket observation.

This code is incredibly hacky, but as long as we are required to do
this from our side, all solutions will be hacky. Please help me make
it as clean as an abomination like this allows.

Note that this fix only works 100% well with Igor's PR in ios-sdk
about message duplication. Otherwise, welcome messages are duplicated.
However you can still see that all duplicated welcome messages are
correctly added after the operator header.

MOB-2183