Apply Unified Customization to Persistent Button
This PR adds Unified customization to Persistent button, together with
default theme. In GVAPersistentButtonVIew the ContainerView was moved to
a stand-alone class so layoutSubviews() could be used properly, since
gradient color can only be applied in that method.

In addition, it adds system message to unified customization
since it wasn't applied before but was actually ready to be used.

MOB-2372