Fix custom cards handling
Sdk differentiate custom card and usual response card based on metadata.
If the type is single choice card and metadata is nil, SDK should draw a message
as a response card.
If the type is single choice card and metadata is not nil, SDK should pass
metadata to renderer and render custom card.

MOB-2342