Cover GVA persistent button with snapshot tests
This PR covers Persistent button with VoiceOver and Layout snapshot tests.
However, due to complications, dynamic font snapshot tests are not working
as expected and will be postponed until later in order to maximise progress
with the project. Rest assured, dynamic font in attributed string works
well in simulator and in real device.

This PR also utilizes textstyle as a key component in making attributed
string scale with dynamic font. Because of that, no default value is
provided in the styles for textStyle anymore.

MetadataWrapper is necessary for making proper metadata json, as the nature
of the Metadata type is somewhat restrictive.

MOB-2375