Update strings in call screens
The changes in this file were done partially. This is because there are some
strings that are using parameters in the string which are replaces on runtime.
These need to be evaluated to see if a backwards compatibility layer is needed
for them or not, which is out of the scope of this task at the moment.

MOB-2538