Update strings to ensure successful snapshot tests
Three strings have changed:

* "we are" instead of "we're".
* Asterisk has the space removed.
* "we will" instead of "we'll"

The changes in which the longer version is used were done because Android
doesn't like quotes on strings, so they had to use escaping or something like
that, but iOS doesn't have that issue, however we would still need to have
escaped strings. Thus we have just removed all quotes. The asterisk space was
removed because it is viewing logic, so it shouldn't be in the string.

MOB-2538