Capitalize accessibility identifier for back button
Before, the accessibility label was "back". Now it is "Back". This is done to
make unification of strings easier, as I don't want to have both strings to say
the same thing. Also, all other accessibility labels start with a capital
letter.

Done in tandem with the snapshot test fix[1].

MOB-2538

[1]: https://github.com/salemove/ios-widgets-snapshots/pull/37