Create backwards compatibility layer for operator name
The idea is that if the string comes from the styling, there is a chance that
the integrator has changed it to use their own string. Thus, the backwards
compatibility layer checks if this is the case by searching for
`{operatorName}`. If it finds it, then it replaces the operator name like
before. However, if it doesn't, then it just uses the operator name directly.
In six months, we can remove all the references to the old `{operatorName}`
behavior, since this is not part of the new strings.

MOB-2507