Internalize templated strings
Instead of creating a layer that verifies if a string has a template or not,
the templates (for example, `{operatorImage}`) have been captured in an
internal extension of the `Localization` struct. That way integrators have a
deprecation that they shouldn't use our strings anymore in case they do, but
also the minimum amount of changes is done to the workings on the SDK. So we can
still use templated strings as right now without going and changing a bunch of
SDK functionality.

MOB-2507