Introduce string providing mechanism
The proposal is that the Core SDK will return to us a dictionary with the keys and
values just as they come from the backend. This is also exactly how the strings
are saved in the `Localizable.strings` file, accessed through the `Localization`
enum. When the Core SDK is configured and is able to provide the dictionary, we
save it to a StringProviding struct. When a string is accessed through the
use of `Localization`, then `Localization` knows what is the key of the string,
and also specifies a fallback in case reading from the file fails. So with this
same key, we can go to the StringProviding, consult the dictionary, and if the
string is not there, use the file fallback, and if not use the string fallback.
This means that with this simple PR, once the Core SDK provides the dictionary,
we support custom locales for the whole Widgets SDK with minimal effort. The
alternative is to create a model which we need to keep in sync between Core and
Widgets, which seems like a big mess.

MOB-2282