Get company name from custom locales
In this commit, support for adding a company name from custom locales is added.
However, if the integrator has set this name through the theme, that one is used
instead of the custom locales. Also, in case the custom locale doesn't have a
string for the company name, but the integrator has specified a company name
through the configuration of the SDK, then that is used. Finally, if neither of
the previous possibilies have been used, then the local fallbacks are used.
Together, the first and the last point mean that there are no backwards
compatibility issues, as if the integrator hasn't set any name, still the
local fallback will be displayed, which is "CompanyName". If they have, then
that is used immediately without caring about configuration or custom locales.

MOB-2686