Improve snapshot test
This PR improves the way snapshot tests are executed. By creating an
extension to UIViewController, we can now create snapshot tests with
less code, time, and effort. And, it is much more readable. Since our
usages of snapshot tests are very limited (for now), this approach serves
those purposes very well.

As this proposal is up for debate, I would like to hear everybody's input
on it. Please share what are your thoughts on making this change.

MOB-2696