Change priority of company name algorithm
Before, the UiTheme took the highest priority in the algorithm to determine the
company name. Now the remote company name has the highest priority, with the
added constraint that it cannot be empty, as this will be the default value on
the default locale. In case it is empty, the UiTheme will come afterwards, then
the SDK configuration, and finally the local fallback.

MOB-2733