Add powered by to leave queue alert
This commit syncs the leave queue alert to how Android works. Also, the end
engagement alert now uses the `showsPoweredBy` variable. This means that it will
now respect the global setting rather than always setting it as yes. This would
bring issues if a client pays for hiding powered by, as it would be wrongly
shown in that case.

MOB-2742