Add support for default values in single choice questions in survey
Default values weren't selected automatically in surveys that showed a single
option. With this change, a new `defaultOption` is added to the survey and used
in case the selected value is not there. Now this works as it should, and how
it works in Android.

The reason why a default option was added and the default value is not just set
as the selected one, is because the question view starts with props, and the
prop is either selected or not. However, the question view wouldn't know if an
option is selected because of user input or because of the default value. If the
view is selected because of user input, then `opt.select(opt)` needs to be
executed only on user interaction, but in case of default value, it needs to be
run automatically as soon as the view is presented. I don't know how to make
this distinction without having the default value separately.

Also, a snapshot test was added to ensure that when there's a default value,
it's selected correctly.

MOB-2747