Fix dependency update lane
There is a workflow in ios-bundle repository called "Post Release" which gets
the version and the checksum from the latest version, sends it to an
`update-dependencies` workflow in ios-sdk-widgets, which then executes a
Fastlane lane that uses that info, updates the templates, and generates the new
files. However for some reason, these version and checksum parameters weren't
used in the calling of the Fastlane lane. Now they are called.

Also, now the replacement of the Core SDK version is done first in the template,
then the template is copied to the root folder, and then it is adjusted for the
latest Widgets version. This way, the root podspec and the  template podspec
always have correct Core version. Before, the template was copied first and then
the replacement of the Core SDK was done. This meant the template was always out
of date.

MOB-2770