Add possibility to change `Podfile.lock` in Post Release workflow
The Post Release workflow in ios-bundle executes code in ios-sdk-widgets, which
sets in place the new version and package checksum. It however omitted the
Podfile.lock, which meant that you still had to manually update it. This PR
changes the version also in the Podfile.lock, while also changing the Glia Core
SDK Podfile checksum.

MOB-2773