Move call visualizer mocks to tests
Some mocks were part of the GliaWidgets, but some of the mocks they use were
declared on the tests. I believe there's some kind of optimization in which
Swift Package Manager adds all files but ignores the ones that are not in the
GliaWidgets/Sources folder. This meant that the mocks in the Widgets didn't find
the mocks on the tests, and on integration, Cortex Financial didn't compile.

MOB-2778