Add updating the widgets version in Cortex Financial when releasing
The new version of the Widgets SDK, which is represented in the Bitrise workflow
as `$NEW_VERSION`, is sent to a script that executes the `update_dependencies`
workflow in Cortex Financial. Then, in the Cortex side, the version of the
widgets will be increased and a new PR made. As this is part of the `deployment`
workflow for the widgets, this means that upon releasing widgets, nothing needs
to be done from our side to update Cortex (apart from approving the PR in
Cortex).

MOB-2783