Add manual locale override capability
The configuration accepts a manual locale override, which is `nil` by default.
If the Core SDK returns an incorrect manual locale override, then the Widgets
SDK also returns an error on configuration to let know the integrator about it.
If the integrator provides a wrong site API key, this commit also makes sure to
inform the integrator instead of hanging.

Also, the settings screen now includes the possibility to add a manual locale
override visually. If it's empty, then it sends nil.

MOB-2798