Do a repo-update on update-dependencies Bitrise workflow
The update-dependencies workflow wasn't finding the latest version of the Core
SDK because the Bitrise Cocoapods cache is updated over the weekend. Now, the
update also does a --repo-update and thus it will pick up the updated version.
This means that messing with the Podfile.lock is not necessary and these
instructions have been deleted from Fastlane.

MOB-2816