Update bitrise in master to latest version in development
Our Bitrise is configured to have the `master` branch as the default branch, so
that is where Bitrise goes to retrieve the `bitrise.yml`. However, our
`development` branch has updates to make everything work well. This commit
brings those changes to master so that Bitrise can pick up these changes. The
reason why everything was working until now is because Bitrise was configured to
be using the `bitrise.yml` saved in Bitrise. However we should configure Bitrise
to use the `bitrise.yml` in the app repository (which points to master).