Create Operator Request Handling Service
There were several place all over the code base where we presented
alerts for media upgrade and screensharing. This however made the code
less maintainable and a new service was made to accommodate all operator
initiated requests. In addition, this service enables us to how Glia
alert in all of the client's app views, not just in ours.

MOB-3040