Address bubble disappearing when tapped during live engagement auth
Make sure that right after Direct ID authentication, the 'maximize' navigation happens when bubble is tapped. The issue was with attempt to present 'GliaViewController' on 'BubbleViewController', which resulted in hidden BubbleView. Now we try to filter out 'BubbleWindow' to avoid presenting on its `topMostController`. In case if this BubbleWindow is the only window, we use 'waiting with timeout' strategy for non-BubbleViewController to become top most and continue navigation with it.

MOB-3226