Add tests for CallVisualizerCoordinator
The delegates inside the coordinators are not tested because each coordinator
itself tests the delegate methods functioning through their own set of tests.
Also, the coordinators in this class are all private, so testing it from outside
would require moving the coordinators to internal properties, which I would say
is not desirable.

MOB-2917