Add unit tests for ScreenSharingCoordinator
The start and only delegate event was tested. In the case of the delegate event,
because it also dismisses the view controller, I attempted to test it too.
However, being an UI operation, it was complicated and flaky to test because
it depended on what was currently present on the screen as part of the testing
app. Several issues with presenting/presented view controllers were encountered,
so I scrapped the plan.

MOB-2916