Proof of concept for handling queues inside of Widgets SDK
A coordinator is put between the start of the engagement and the root
coordinator, which shows the media selection screen. After the media type is
selected, this information is sent back to the root coordinator. Currently in
some circumstances, the media selector doesn't go away, but this is because a
cleaner approach should be done, in which the media selector goes to the root
coordinator instead of sending back the information to the Glia class. Or, the
root coordinator should be in charge of showing the media selector. In any case,
it's not complicated to achieve.

What media is shown is handled by an array to which you can send the media types
and the queues. The integrators would then just need to send us this info on the
`start` method, and we would be the ones who use this to start the engagement.

MOB-3298