Fix crash on onHold during LO
When the visitor is put on-hold during LO, the crash occurs, because `OnHoldOverlayVisualEffectView.draw` method is called on background thread by `layer?.render(in:)` in `WindowCapturer.native`. Since we need to keep screen capturing on background thread, `draw` method execution was wrapped into main queue. This issue is related only to native implementation for WindowCapturer, and is not reproducible for flutter side.

MOB-3332