Add subset of new SwiftLint rules
Add the following rules to the codebase:

- multiline_arguments_brackets
- multiline_literal_brackets
- multiline_parameters
- multiline_parameters_brackets

The main point is that sometimes it comes up in our PRs where should parenthesis
start, end, how to delimit parameters of a method, etc. These rules in essence
say that if you decide to put parameters in different lines, then the starting
delimiter should be with the name of the method, and then also the end one
should be on its own line. Also in this case, all parameters should go in their
own lines. These rules do NOT mandate that you must put parameters into their
own lines always, however. You are free to put them all in the same line if you
want.

As you see with the changes, they all make the code look more like the structure
and the readability we strive for, so I hope that adding these rules will make
sense to you. Otherwise, you can comment in the PR.

MOB-3175