Add subset of new SwiftLint rules
Add the following rules to the codebase:

- unowned_variable_capture
- vertical_whitespace_opening_braces
- vertical_whitespace_closing_braces

These are mostly to end the inconsistency we have where some parts have spaces
after opening braces, and some don't. We decided some time ago that we wouldn't
but nothing was enforcing it other than PRs.

This is the last set of rules that are going to be added.

MOB-3175