Create Alert Manager
Alert Manager brings everything related to alerts to one place. With a
single interface, every alert will be handled by present() method.
An alert can be presented in a root view, or in a separate keyWindow,
depending on use case.

This PR is a partially complete product, because early feedback is necessary,
before moving forward with it. Currently only global alerts are handled,
and alerts in root view have been left out for now.

MOB-3373