Hide EntryWidget sheet before engagement
Due to the different nature of chat and call views, the behaviour of
alert dismissal worked way different. For Audio and Video engagements,
when confrmation dialog was declined, the engagement view was dismissed
alongside Entrywidget sheet. This was not intended. But for Chat, the dialog
dismissal didn't dismiss EntryWidget sheet. To align them all, the entry
widget is not dismissed when media type has been selected.

It also looks much better visually now with the animation.

MOB-3760