Add logs and telemetry for Entry Widget
This PR adds logs to EntryWidget, and aligning iOS with Android logs.
Because logger was added as a dependency to QueueMonitor, it was necessary
to move queueMonitor from Glia environment to Glia's class property due to
Logger being initialized during GliaSDK initialization. Because of that
some changes can be seen in tests as well.

MOB-3481