Do not cancel enqueuing if new enqueue is the same engagement type
Currently if there visitor is enqueuing and decides to initiate engagement
again, then that will cancel their queue ticket and create a new one. This is
really bad from UX perspective because a person would potentially lose
a great spot in a long queue, even though their reason was to just resume the existing one.

The new solution would check for the engagement kind and if the current one matches
the new one, then instead of cancelling the old, the old one is resumed.

There were technical difficulties to achieve that because currently the SDK converts EngagementKind
to simplified mediaType, which makes this upgrade impossible. This PR simplifies the approach significantly
and allows the SDK to rely on EngagementKind as long as possible and only use mediaType when absolutely needed.
This PR lays a pathway forward to continue removing mediaType from places where EngagementKind is suffice.

MOB-3931