Make interactor as published property
If integrator creates EntryWidget and then re-configures the SDK, EntryWidgets stops handling engagement states, because it owns outdated interactor instance.
To solve this problem we need to re-subscribe on interactor.currentEngagement property once the SDK is reconfigured. This PR does that by making interactor as a published property. EntryWidget will always be up to date with the most recent interactor.

MOB-3965