Fix snackbar bottom paddings
This PR does the following:
1. Initial padding for Snackbar is correct in the chat
2. Initial padding for Snackbar is correct in the call
3. Snackbar is decoupled from keyboard actions entirely
4. Snackbar has been moved to the new DI
5. Snackbar is stored as pending, if chatview is not in hierarchy. This is needed
When a persistent snackbar is shown, such as connectivity issue snackbar, but during
audio call user hasn't used chat yet. This logic allows to show the persistent snackbar
in chat later. NOTE: If chat was upgraded to audio, and connection issue starts, this flow
is skipped entirely.

MOB-4961